♻️[Refactor]#9 - 지원자 상세모달 리팩토링
ModalViewModel로 로직 분리
ModalViewModel -> ManagementViewModel 단방향 참조로 로직 재활용
