+++ Muje/Muje/Common/Enum/ApplicantModal/ButtonType.swift
@@ -1,0 +1,13 @@
+//
+//  ButtonType.swift
+//  Muje
+//
+//  Created by 조재훈 on 8/18/25.
+//
+
+import Foundation
+
+enum ButtonType {
+  case left
+  case right
+}
